Split the node to Assign and Set.
