+++ index.md
@@ -326,6 +326,7 @@
 ## Objects
     SvVertexGroupNodeMK2
     SvVertexColorNodeMK3
+    SvAssignMaterialListNode
     SvMaterialIndexNode
 
 ## Layout
